requirements.txt
torch>=2.0.0
pipecat-ai[silero,daily,openai,deepgram,elevenlabs,twilio]==0.0.63
aiohttp>=3.9.4
torchaudio>=2.3.0
channels>=4.0.0
requests==2.32.2
twilio>=9.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
loguru>=0.7.0

# Compatibility and integrations
pydantic>=2.8.0,<3.0
pydantic-settings>=2.2
openai[realtime]>=1.68.2
livekit-plugins-openai==1.0.18